+++ river/tree/hoeffding_tree.py
@@ -58,7 +58,7 @@ class HoeffdingTree(ABC):
         self,
         max_depth: int = None,
         binary_split: bool = False,
-        max_size: int = 100,
+        max_size: float = 100.0,
         memory_estimate_period: int = 1000000,
         stop_mem_management: bool = False,
         remove_poor_attrs: bool = False,
